Controller templating and results table gen.
Static controllers were replaced by controller templates and a generator script.
This allows easier testing of multiple parameters.

A Python script was added to the report folder, in order to automatically
generate the results partial table in LaTeX automatically. This also allows for
later updates of those values.
Note: the LaTeX syntax of the output was not verified.